requirements.txt
PyYAML==3.11
Flask==0.12
redis==2.10.5
pytest==3.1.0
flask-sqlalchemy
mysql-python
